requirements.txt
Flask==2.3.3
tensorflow==2.13.0
Pillow==10.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
numpy==1.24.3
